+++ example/config/initializers/config_good_migrations.rb
@@ -1,1 +1,0 @@
-GoodMigrations.config.permit_autoloading_before_date = ENV["PERMIT_BEFORE_DATE"]