+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-* Add `permit_autoloading_before_date` configuration
+* Add `permit_autoloading_before` configuration
 
 ## 0.1.0